PiCamera/requirements.txt
AWSIoTPythonSDK==1.3.1
certifi==2018.4.16
chardet==3.0.4
futures==3.2.0
idna==2.6
PyYAML==3.12
requests==2.18.4
tinys3==0.1.12
urllib3==1.22
